requirements.txt
appdirs==1.4.3
boto3==1.4.4
botocore==1.5.37
configparser==3.5.0
docutils==0.13.1
jmespath==0.9.2
packaging==16.8
pyparsing==2.2.0
python-dateutil==2.6.0
python-magic==0.4.13
requests==2.13.0
s3transfer==0.1.10
six==1.10.0
